=General *Now supports 65k item ID's. -Technically the max is now 65535 but 65k sounds cooler. -This is confirmed to work with the 2018-06-21aRagexeRE and should work with older -clients. Be sure to apply the SQL update 3CeAM_svn876.sql. -Make sure to backup your server before running.
*Now saves guild storage permissions.
-It only alows setting it but its not yet coded to set if a NPC allows acces or
-not. This will be coded in a future update.



=SQL
*Updated main.sql to allow cards to work in the 65k range.

*Added the 3CeAM_svn876.sql file to update card slot entrys to support cards in
-the 65k range.



=Skills
*GD_GUILD_STORAGE
-Skill Tree Requirements updated.
-This is the final time its being updated since I now have official info on
-its requirements.

git-svn-id: https://svn.code.sf.net/p/v1-3ceam/code/trunk@876 cad27aaa-dce3-4a30-a00a-e4fd67c11881

+++ rewrite/src/common/mmo.h
@@ -226,13 +226,13 @@ struct quest {
 
 struct item {
 	int id;
-	short nameid;
+	int nameid;
 	short amount;
 	unsigned int equip; // location(s) where item is equipped (using enum equip_pos for bitmasking)
 	char identify;
 	char refine;
 	char attribute;
-	short card[MAX_SLOTS];
+	int card[MAX_SLOTS];
 	unsigned int expire_time;
 };
