+++ rewrite/src/map/searchstore.c
@@ -382,7 +382,7 @@ void searchstore_clearremote(struct map_session_data* sd)
 
 
 /// receives results from a store-specific callback
-bool searchstore_result(struct map_session_data* sd, int store_id, int account_id, const char* store_name, unsigned short nameid, unsigned short amount, unsigned int price, const short* card, unsigned char refine)
+bool searchstore_result(struct map_session_data* sd, int store_id, int account_id, const char* store_name, unsigned int nameid, unsigned short amount, unsigned int price, const int* card, unsigned char refine)
 {
 	struct s_search_store_info_item* ssitem;
